Switch to FIRST_BYTE() logic for checking REBNOD flags
The rules for strict aliasing are a bit nebulous, but one thing is
known: that `char*` and `unsigned char*` access defeats it.  This
means that if a void pointer or node pointer is in one's hand and
one is interested in knowing what subclass it was assigned under,
the most foolproof way to make that determination is through checking
the bits via bytes.

IS_END() was converted to use byte access because it was being used
with things that did not have full Reb_Header-sized data on it (the
shortcut rebEND).  But since cells and series both had a common
Reb_Header union at the beginning, it was not always followed for
finding out if an assignment had been through REBSER* or REBCEL*.

This attempts to move some cases toward doing the discernment of the
assignment type via byte access of the node header only.