Allow subdirectories in %src/core
The original layout of Rebol was one flat directory with files whose
first letter indicated the subsystem that it was part of, followed by
a hyphen, and the rest of the name.

Directories offer an advantage for helping those browsing a project
get a sense of the structure.  In particular, in the modern age, sites
like GitHub let you put a README.md in each directory which lets you
explain why those files are grouped together.

This advantage for those browsing a codebase is significant enough to
outweigh the questionable advantage of not using subdirectories.