Separate dumping value properties from crashing
The concept of being able to panic(value) included printing some
useful diagnostics about that value.  This information could be useful
in a debug build without tying that into also crashing.

This commit separates the code, and introduces a new DIAGNOSE native
that will output information about a value and its internals.