Fix PARSE of string binary alias, eliminate SER_LEN()
Historically in Rebol, the "length" as conceived by a user of a series
being indexed into a node was always measured in terms of the "width"
in units of that series.  Hence extracting a series node and asking
for its "length" would give the same anwer as asking a value holding
that series node its length.

With UTF-8 Everywhere, this idea was no longer true.  A particular
issue is that while you can tell if a series is a string (and thus has
a cache of its length in codepoints in the MISC slot)...that series
could be aliased AS a BINARY! which would not want to use it in this
way.  So if you have an operation like PARSE which operates on both
BINARY! and ANY-STRING! values...you have to ask the question of the
*value* what the length is, and not the extracted series...to get
the right point of view.

This makes that change, and eliminates the misleading SER_LEN()
function, which could almost never be used correctly.  PARSE callsites
were changed to use VAL_LEN_HEAD() to get the right interpretation,
and other callsites to use BIN_LEN() or SER_USED().

This corrected a bug in length checking which was falsely thinking that
some string matches in BINARY! series that were string-aliases were
out of bounds, as the index position in the binary was past the length
in codepoints.

Also: to get the proper sensitivity to apply for find operations, the
value-based "Find_In_Any_Sequence()" return was adapted to be used
pervasively in PARSE.  This allowed eliminating redundant code that
decides to include the delimiters in TAG! in a parse, as that decision
is already made by the shared find code at this point.