"Modularize" the WebAssembly build
By default, Emscripten creates an object called "Module" which holds
the Wasm heap, exported C-compiled functions, and helper routines.
But if all emscripten codebases did this, you wouldn't be able to load
two codebases in the same page because they would overwrite each other.

This takes advantage of the "MODULARIZE=1" option in Emscripten, to
name a function which gives a Promise back to make a module.  This is
then poked away inside the `reb` object, as `reb.m`.

Also, this goes ahead and wraps up some miscellaneous global tables and
puts them inside the `reb` object as well.