FFI compilation patches
The FFI was close to working, but had atrophied a slight bit.  This
brings it back to working condition for the demos: gtk.r, qsort.r,
qsort_r, printf.r, varargs.r.

Problems were:

* Did not work with out-of-source builds due to %make-spec.r using
  a relative path (%../extensions) to find the library and vector
  extensions.

* Failed to terminate binaries with a zero byte.  All BINARY! must
  a-priori terminate with zero at this point, in order to be able to
  be aliased AS TEXT! if desired.  Restructured to make termination
  more practical during the aligning buffer store creation.

* "Rescued" routines now cannot return void*, they must always return a
  REBVAL*, even if this pointer is ignored by the caller.

Included in this change is removal of several FFI-specific terms from
the %words.r list.  Not all are removed, but the terms related to the
ABI are used rarely enough that a libRebol API solution is sufficient.