Switch FAIL back to blaming a QUOTED! word/path
When the @word type was first created, its applications were not
fully understood.  It seemed to look nice to be able to have a
skippable location for where to report failures use the @ symbol,
asuming that the @ would come to be connoted with "where it's at",
e.g. some kind of location.

But-reducing @-types began to take a different direction, implying
a way to implement modal parameters, deferred execution branches, and
likely being the means of expressing "datatype!" and maybe "typeset!".

This returns to using `fail 'foo [...]` syntax instead of the
`fail @foo [...]` syntax, for ascribing blame to a variable as the
cause of an error.