Fix ECDHE spelling errors, add mod-crypt README.md
Perhaps due to looking more pronounceable, there were several wrong
spellings of ECDHE as ECHDE.  Not all were in comments--so could be
a source of problems when the values didn't match when they should.

This fixes that, and writes up a README.md to sum up the current state
of affairs in the Crypt module.