Permit module source to be ".cpp" as well as ".c"
The draft implementation of the extension model had some naming systems
based on a different idea, where extensions might contain many
different modules.  This seemed to be jumping the gun, and had a lot
of difficult properties due to that...so Ren-C switched this to being
one module per extension.

In that process, some manual name generation for files was changed to
be automatic.  It is a bit crufty--having been based on a partial
implementation of a more ambitious idea.  However, it was less code
and was an improvement at the time.

The auto-generated naming didn't do much checking and was just a
search and replace.  This makes it more rigorous using PARSE rules
that enforce the name scheme (even if it's kind of a poor one), and
expands it so that .cpp files are supported--in case the extension
is written in C++.