Allow SET to take VOID!, remove /OPT refinement and SET*
The current mode of thinking in Ren-C is that SET, SET-WORD!, SET-PATH!
do not complain when assigned either NULL or VOID!.  This is based
on the premise that the main channel for catching typos or errors is
through WORD! access...which complains when a VOID! is referenced.
This catches typos on the GET side.

(Catching typos on the set side with values isn't really possible,
as `mispeled-var: 100` involves no void!s or NULLs anyway)

This brings master up to date with these ideas, including ones discussed
today in chat.  That is to make SET have parity with SET-WORD! and
not require any /OPT or /ANY refinement.  Instead, it's up to users
to check the value they are assigning.  Examples would be:

    set var non void! (...some expression...)
    set var non null (...some expression...)
    set numeric-var ensure [integer! decimal!] (...some expression...)

This also adds a trivial version of NON that covers these cases, to
be expanded upon later.