Fix error message truncation in rebError_OS
Modern GCC has a warning that notices when you use `strncpy` and do
a length-limited copy into a buffer, where the length limit is the
size of the buffer exactly.  This means your truncated string may
not be terminated.  That was the case in the truncation of POSIX
error messages.  This drops the length by 1 and ensures that the
last element of the buffer is a NUL terminator.