GET-WORD! errors on VOID!, only suppresses ACTION!
This change aligns Ren-C with Rebol2, in considering the VOID! (e.g.
"UNSET!") state to be an error when fetched via GET-WORD!.  That means
that GET/ANY must be used to read VOID!.

The motivation for this is to make sure that when people use GET-WORD!
to suppress function evaluation, they aren't also getting something
they may not want...e.g. the suppression of errors on typos if the
word looks up to void.

As a compromise for performance, GET* is added to mean "GET/ANY" as
a native.  This makes `get* 'var` a bit closer to the performance
profile of `:var` because there is no path processing...and it is
a native so it does not incur specialization overhead.

In light of this convention, the loaded term "Get_Var" is substituted
with "Lookup_Word" for cases that are agnostic about VOID! and given
direct (movable) pointers into contexts, with `Get_Word` repurposed to
mean "move value into cell and ensure it is not void".  Then the
oddly intimidating names like `Get_Opt_Polymorphic` are simplified to
just `Get_Var` to imply taking either a WORD! or a PATH! as a way of
finding a variable slot.