Hack for Unicode keyup paste bug in Windows Terminal
There is a problem with certain high unicode characters when using the
ENABLE_PROCESSED_INPUT option with the console, when a paste command
is run (either through Ctrl-V or the window menu's Paste).  Only
key up events are sent for some characters.

This integrates a patch from libuv for the issue.  It also makes a
shorthand for accessing the key_event union member to make the code
a little bit more readable.