Permit matching of WORD!s in strings
This adapts the code powering FIND to be able to run FINDs against
WORD!s in strings, which is a feature in Rebol2, R3-Alpha, and Red:

    >> find "abchellodef" 'hello
    == "hellodef

This then feeds into being able to do such matches in PARSE as well.
Rebol2, R3-Alpha, and Red do not do this:

    >> did parse "abc" ['ab 'c]
    == #[true]

It also supports finding integers in strings, which is a feature of
Rebol2 and R3-Alpha but not Red:

    >> find "abc10def" 10
    == "10def"

    >> did parse "abc10" ['a 'bc '10]
    == #[true]

This required making the FIND infrastructure able to operate on "cells"
which are a view on values that neglect their quotes, and only focus
on the actual cell structure.