Use GET-BLOCK! with KEEP to mean DO, not REDUCE
When GET-BLOCK! was first invented, it was thought that it might be a
convenience for REDUCE:

    >> append [a b c] reduce [5 + 5 13 + 7]
    == [a b c 10 20]

    >> append [a b c] :[5 + 5 13 + 7]
    == [a b c 10 20]

This idea fell from favor.  It's not really adding significant
expressivity to the language, and it would only work with literal
blocks (e.g. `:block` would not reduce that block).  Other more
interesting usages for GET-BLOCK! that solve foundational language
problems arose.

Prior to the new thinking, PARSE used `:[...]` to mean "reduce this
block as data", e.g. to make KEEP be able to accept material that did
not originate from the input series at all.  This is still a necessary
idea, and GET-BLOCK! may or may not be how to achieve it.  But thinking
that it needs to REDUCE instead of simply DO is less general...as
one could say **:[reduce [...]]** in order to get that effect.

This switches it to DO semantics, and also makes GET-BLOCK! inert
in the evaluator for now.