Fix PARSE END in TO and THRU, APPEND of NUL byte
An attempt to formalize the idea of '\0' not being addable to series
led to stricter checks on characters, e.g. that UP_CASE() and LO_CASE()
inside the system would not accept 0.

This led to a shortening of the enumeration range of TO THRU rule
processing.  It didn't step over the end of strings.  However, that
broke the matching of END rules when used with TO and THRU if those
rules were in a BLOCK! (e.g. `to ["a" | end]`), because it wasn't
willing to test on the end position.

This fixes that issue, and adds a protection to APPEND/CHANGE/INSERT
to disallow the addition of '\0' bytes to strings via CHAR! or an
arbitrary BINARY!.