Dropped character bugfix for windows console
With the addition of processing for characters which receive a key-up
with no key-down, there was a problem that the key-ups had previously
been consumed implicitly by the key-down code.  But when a key-up was
treated as a key-down, this implicit ignore then consumed the next
key-down.

The consumption of the key-ups was actually a bug, just one that went
unnoticed when all key ups were ignored.  This resolves that bug by
letting the tail of the loop consume the key-down (as it does with
other event-generating cases).