AUGMENT patch for working with specializations
AUGMENT made actions with more arguments than their underlying
functions.  The system came to tolerate this in the asserts--checking
only that derivations had as many or more elements in the paramlist
than those they derived from.

Specializations--however--currently require a frame long enough to
increment all the way through for the level that's running.  This
means that an exemplar frame with fewer elements cannot be used
(unless that rule changes--it's currently done for optimization so
that special doesn't need to be checked for END).

Thi makes a patch so that it will at least work for now, if you
AUGMENT a specialized function.