Fix environment variable reads with MODULARIZE=1
Emscripten contains a simulation of the environment for getenv() and
setenv().  Previous to modularization, it was possible to set this
up via an ENV global variable.  It got more picky with MODULARIZE=1,
where you have to set it up at a very particular moment in preRun(),
and you also have to export it in the link command line.