Fail on TO TEXT! MAKE CHAR! 0
Previously TO conversions for ANY-STRING! wouldn't check for the 0
codepoint.  This adds a check for that, and takes care of a few
edge cases in PARSE where the internal terminator of a BINARY! was
being recognized as if there was an actual zero byte at the end
of the meaningful user data.