Switch to using Python3 for Firefox Scripting
Deprecation warnings and malfunctions of the "six" compatibility
layer for Python2 => Python3 makes it seem that switching to Python3
is something that should be done.

It seems that the marionette driver and script will work unaltered.
However there is a problem about JSON decoding if older Python 3.5
is used, so this switches both web builds to Xenial.