Attempted improvement of flushing of input after Ctrl-C
This does a small theoretical improvement on flushing keyboard input
if a Ctrl-C is encountered...but it only works if the moment of
cancellation is during console entry.  If pending input is accrued
while interpreter code is running, then it the halt won't be received
in the place where the keyboard buffer can be cleared.

To do what is actually desired, then the keyboard clearing code needs
to be able to get a hook into the Ctrl-C handler, or provide the Ctrl-C
handling service itself.