Fix parentheses in escaped codepoints
Something in the UTF-8 everywhere conversion caused escaped codepoints
like `^(FFFE)` to come out as `^"FFFE)` in mold.

This remedies that issue, as well as adding a test to confirm that
the unicode "byte order mark" is not filtered out in string reads.
(It might be better to consider them an error on plain reads of text,
as with CR.)