Update %make-zlib.r import, adjust comments, reorg
Comments in the %make-zlib.r file suggested that it had last been run
on a 2013 version.  Several bugs were fixed in a 2017 version, and as
the script hadn't been run recently to digest the zlib sources into one
header and one source file...it had a few things that needed fixing.

This commit fixes those issues.  But as it turned out, Shixin had
updated Zlib (and just not updated the comments).  So the version that
was in use was already the 2017 version...just with outdated copyright
notice and version info.

So long as the script was being updated, this also changes the comments
in the auto-generated file to the `/* */` style...which is a convention
used in Ren-C to help notice when one is editing an auto-generated
file.  It also puts DEFLATE and INFLATE natives into the %u-compress.c
file (as they were in %n-strings.c for some reason).  Light review
of %u-compress.c -- which is pretty well vetted -- merely updates it
to the "two spaces off for line comments" convention.