Don't auto-"..." limited molds, return truncation flag
The MOLD/LIMIT option allowed you to provide a limit of how many
characters of a mold were desired.  But this limit had to be at least
3, otherwise there wouldn't be enough room for an implicitly inserted
"..." on truncation.  An assert would fire if users passed in too
small a limit.

Rather than firing an assert, this switches to not including the "..."
in the truncated output.  Instead, callers can ask for a flag saying
whether there was a truncation or not, and add in their own handling
(e.g. perhaps an "expand" button which would call for a larger mold
in a GUI environment).

Also fixes a related bug that was incorrectly comparing the length of
the total mold buffer to the limit--instead of just the currently
pushed mold.