Use signal.h instead of sys/signal.h
Trying to build against `musl` led it to complain about the use of
`sys/signal.h` being forwarded to signal.h.  It seems to work either
way, as Linux `signal.h` generall appears to include `sys/signal.h`.

If this turns out to be a problem, then a compile-time #define may
be necessary.