Take SECURE startup code out of main init
The SECURE concept was not fleshed out particularly well in R3-Alpha
(and in fact, disabled).  It was moved into a separate module and
taken out of builds like the Web build.  However there was some
residual commented out code in the main startup related to it.  This
moves that into the mothballed SECURE module so it's out of the way,
but can be contemplated if and when the SECURE module is revisited
in the future.