Fix PARSE case handling with BITSET!
When UTF-8 Everywhere was introduced, it became possible to match
ANY-STRING! against BINARY!.  So instead of marking all binary parses
as being case insensitive, the question of what one was matching
against came into play.

This resulted in confusion over the behavior of BITSET!, which may or
may not be intended as a character-like "charset".  The traditional
interpretation for BINARY! is that it's supposed to be literal bytes.
However, they were being case-matched as CHAR!, which was not only
a break with history but also has the problem that situations which
were matching 0 bytes were asserting due to the limitation on not
trying to canonize case on codepoint 0.