Fix missing BYTE in Windows Encap decoding
Because Encapping isn't being used right now, it went unnoticed that
the definition of BYTE went missing in the ENBIN/DEBIN change.  On
a BINARY! the definition is simply SKIP, so this changes to use that
(instead of replacing it with the BITSET! of all values that was
used previously).