Probe() helper, mold NULL as `; null`
The PROBE() macro cannot be called from the C debugger watchlist, and
the underlying routine Probe_Core_Debug() takes 4 parameters and is
awkward.  This makes a single-arity `Probe()` that is a non-macro
variant, specifically for calling from the C debugger.

Also molds NULL values in PROBE() as `; null` instead of nothing.

To make it easier to Probe() NEVERNULLs, this renames the member as
simply `p`.