Extend backdated GLIBC to cover pow() and log()
Due to GLIBC's general tendency to introduce "improvements" that
cannot be opted out of, version 2.29 added sped up versions of math
functions.  If a newer compiler is used then it will bind to these
versions, making the executable unable to run against an older glibc.

A similar problem was addressed with symbol remapping and wrapping
when this happened with fcntl().  This expands the hack that was done
to permit that to the optional build environment variable
USE_BACKDATED_GLIBC.

So before the make.r is being run, say:

    export USE_BACKDATED_GLIBC=1