Standard C warning compliance for newer gcc
The C standard checking got stricter, now warning that abs() is
truncating a `long int` as an int.

Also the function type checking on TRASH_CFUNC does not accept a
generically typed function pointer in assignments.  This feature was
not used in the code, but is used in the stackless branch.