Stop TAKE corrupting binary-aliased UTF-8
It is possible to create a BINARY! view of an ANY-STRING! using the
AS BINARY! operation.  These views are mutable, but the mutations
are not allowed to make the UTF-8 invalid.

Enforcement of this policy wasn't pervasive in the initial code for
UTF-8 Everywhere.  This advances it a bit by making the entry points
for TAKE call through Modify_String_Or_Binary(), and then enhances
that routine to handle more cases than it did prior.