Fix evaluator temporary TOOK_HOLD behavior
A misguided change which moved TOOK_HOLD to be a feature of feeds
instead of frames was made.  The lack of tests on errors for modifying
code that is running meant this went unnoticed.

This is a quick fix of the non-stackless branch.  Since frame flags
are somewhat saturated here, it takes the DEBUG frame flag away--which
was used to ensure EVAL_MASK_DEFAULT was |'d in with any flags.  Since
that feature was for forward-thinking future use only, and it's done
in a totally different way in stackless, it's safe to remove.

Two simple tests are added to show basic operation, but more should be
written when time permits.