Tidy BINARY! and ANY-STRING! path set mechanics
The binary path setting branched out from the string path setting, so
it had the same code.  Both are changed to disallow the path poking
of a binstr into a binstr by defining that as "taking whatever the
character AT the string was"...since that makes little sense and
could be done as **str/(n): first other-str** more clearly.

It also stops zero byte poking into strings.