Kill off IS_INERT_KIND
The quick test for whether a kind is inert or not is based on an
ordering trick of the datatypes.  All such tricks are supposed to be
confined to %sys-ordered.h, so that they are vetted whenever the
%types.r table is updated.  However, IS_INERT_KIND() was not in the
%sys-ordered.h file.

When it was looked at to move it there, it turns out there already was
a better-named test ANY_INERT_KIND()...and IS_INERT_KIND() was not
actually being used.  So this just deletes it.