Eliminate ASSERT_NOT_END()
The "ASSERT_NOT_END()" macro was introduced because some points at
the code wanted to do a faster test that didn't check for cell validity
in the cell case.  The fact that its existence is due to being faster
isn't very obvious at the callsites, and since all it's doing is
calling KIND_BYTE_UNCHECKED() then saying that explicitly is more
clear in practice.

(Anything that helps reduce the overall number of useless declarations
in favor of making the source more clear is good.)