Patch single-value limiting code in mold
An inlining of a dying function into the one place it was used led
to some convoluted code, that apparently had a bug.  This simplifies
that code to be more obvious--though the feature should be reviewed
in light of making it possible to make a single value mold at a more
limited length (vs. the mold overall).