Homogenize comparison CT_XXX functions
There was a somewhat confusing mixture of parameterizations and return
values for the "extensible comparison function" interface of R3-Alpha.
While some routines returned the familiar "-1 0 1" for less, equal,
and greater...there was a seeming attempt to optimize by passing in
which operation was desired.  e.g. this way if only equality were to
be tested, then it might have a short cut.

The confusingness of this parameterization (via magic numbers) also
mixed up with the notion of "strictness", as well as the intrinsic
truth that many types do not naturally compare with each other (and
some aren't ordered even within their own typeclass).  As a result, a
spectrum of unpredictable routines emerged.

Python 3 seems to have approached this problem with a reasonable
solution, with "rich comparison":

https://www.python.org/dev/peps/pep-0207/

This replaces the idea that all types are intrinsically comparable
just to make SORT of any data collection have *some* result.  Instead,
users are encouraged to define their own comparison operator for
collections of mixed types.  Also, each operator can be optimized
independently... so if someone is only interested in testing for
equality then the specifics of a less than or greater than test can
be avoided.

(R3-Alpha's agenda seems more like Python2, which made all ordered
comparisons give an answer--even if via some semi-random mechanic based
on internal properties like pointer value or enumeration constant
ordering).

It will take significant work to implement a better defined idea, but
simplifying the hodgepodge to the -1 0 1 model across the board makes
it easier to cohere any model.