Make SELF hidden in bootstrap object
As a bootstrap workaround, a system subobject is created manually by
code to use before the system object is available.  There was a test
to make sure this object was equivalent that was mistakenly passing
because it was believed the test returned `0` for equality, when
that actually meant not equal.  The aspect making the objects unequal
was that the SELF field was not hidden in the bootstrap object.