Use enum Reb_Symbol in compression APIs
The REBSYM datatype is overloaded in the C++ build to do some safety
checking to help avoid accidentally believing two ANY-WORD! spellings
are the same just because they have no SYM_XXX in the enumeration
(and hence would return SYM_0 even if different spellings).  But C++
types are not geared up to be exported in the %sys-core.h API list...
in case a C build tries to mix with a C++ one.

For this reason, the REBSYM was not used in specifying the type of
compression to use...even though it is helpful to be symbol-based.
Instead REBSTR* was used.  Not only is this more inefficient, but it
also couldn't be used during bootstrap--a time when compression has
to be able to be called.

So `nullptr` was used to mean "assume REBSTR* would be Gzip".  This
complex solution had to be documented, and really the simpler thing
would be to just use the `enum Reb_Symbol` enumeration instead of
the wrapper REBSYM.

This makes that change.