Fix ADLER32 hash, make /PART work on strings
The ADLER32 hash was giving incorrect results because the algorithm
was being primed with `0L`.  This is suggested by the zlib docs, but
the actual correct value to start with is `1L`.

In addition to resolving that issue, this goes ahead and makes the
/PART option work for checksums of strings.  In the early days of
UTF-8 everywhere it introduced a complication and wasn't bothered with,
but now that things have matured there is a facility for it.