strsize replaces LEN_BYTES, kill old byte helpers
R3-Alpha had some byte-oriented aliases for functions that classically
operated on `char*` but not `unsigned char*`.  So LEN_BYTES() was a
casting alias for strlen().  This was embraced in early efforts to
make it const-correct...so there were more routines like COPY_BYTES()
or APPEND_BYTES_LIMIT().

Most of the callers of these routines disappeared over time, as things
like the platform-specific character type were eliminated.  This left
most of these routines unused.  So this commit gets rid of them.

However, the term of "length" has a special meaning in UTF8-Everywhere
to mean the *number of codepoints* in a string.  While "size" counts
the number of bytes (in a size_t sense).  This makes strlen() something
of a misnomer, as it counts how many bytes there are until it reaches
a zero in a sequence of bytes.

This subsumes the LEN_BYTES() function into an alias for strlen()
called `strsize`.  It accepts -either- signed or unsigned characters,
but uses C++ templates in the C++ build to make sure that other types
aren't passed accidentally due to the cast.  The name crosses over to
look very much like strlen() the primitive (because that's all it is
doing), while still conveying the point that it's a number of bytes
and not a number of codepoints.