Make rebQ/rebU act as rebQ1/rebU1 in C89 build
Most calls to rebQ and rebU are single arity, but required the rebQ1
or rebU1 variants in C89 to dodge the need to supply a rebEND.  This
was a source of frequent errors as it's very easy to forget...and
at runtime can be confusing.

This switches it around so that rebQ and rebU are single-arity in the
when REBOL_EXPLICIT_END is defined (e.g. C89 or C89-compatible source).
Access to the full versions are avilable through the longer-named
rebQUOTING and rebUNQUOTING, but those would require rebEND.