Remove va_list traversal during GC
During GC, va_lists were "reified" in order to turn them into arrays,
because using ordinary traversal to mark their elements would exhaust
them in a way that could not be restarted.

This was done early in the development of allowing execution of va_list
and there are two things wrong wih it:

1. It's actually possible to make a copy with `va_copy` of the state of
   a va_list at a particular point of enumeration, and enumerate that.
   Thus there's nothing forcing reification of the va_list in a GC
   if one didn't want to.

2. The items passed to the variadic are already expected to be GC safe.
   They're either API handles, ARG()s in frames, or instructions that
   have been explicitly delegated to be freed by va_list traversal.

Review of va_list feeds may deem them not worth the code complexity for
handling them.  Values must be reified into `fetched` cells if they
have a rebQ() or if they might be needed for lookback.  Error messages
are not able to provide context.  Forming a temporary array may be
the more sane approach...this is still being considered.