Performance tweak: lazy initialize binder for API calls
In order to be able to execute the UTF-8 text portions of API calls,
the scanner has to be complicit in the binding.  As it splices in
values in the middle of text runs at arbitrary depth, the bindings
on those spliced values should be left as-is, and the text bound.
(Making matters more complex, the binding environment varies based
on which module the last native call was in...e.g. the module is
the main context.)

Initializing and shutting down binders is relatively expensive, and
hits very sparse cache locations.  This improves matters a bit by
only building the binders when WORD!s are scanned.  This means calls
like these would not need to set up binders:

    REBVAL *block = rebValue("[", item1, item2, "]");
    rebElide(some_func_taking_integer, "10");