Add missing rebEND terminals caught by C++ build
The unpacking variadic template interface of C++ is able to distinguish
the terminal condition in a recursion.  When it is handling the last
parameter it looks to see if REBOL_EXPLICIT_END is in effect.  If it
is, then it makes sure that last slot matches the type of rebEND.

(Note: currently rebEND is a `const void*`, because it must be a type
that has the same alignment as a character, since it is using a
2-character byte string as its implementation.  This would conflate
with any `const void*`.  Another possibility would be to export a
custom type for rebEND and point to the global END_NODE...but the
short character form was chosen for simplicity, so this should be
thought about.)