Add C++ support for varaidic parameter unpacking
The only mechanism C has for variadic function invocation is the
`va_list`.  libRebol works around this limitation to provide a usable
API that allows the mixture of code fragments and existing values to
splice...though this method is limited to only allowing pointers to
specifically fabricated arguments.  Hence this is not legal:

    REBVAL *ten = rebValue("add", 4, 6);

One desire with C++ is to be able to pass fundamental or custom types
in this way.  Since `va_list` is not something that can be built up
programmatically (in a cross-platorm way), this requires use of another
mechanism.

As of C++11, it is possible to write variadic functions which
recursively deconstruct their arguments at compile-time with templates.
This can be used to build a fixed-size array of pointers to arguments
on the stack which is isomorphic to the (usually stack-based) va_list.

This introduces alternate definitions of functions like `rebValue`
which are based on variadic unpacking.  Instead of passing a head
item pointer as `p` and a `va_list` for the remaining arguments, it
builds a call to the same function with the `va_list` argument as
nullptr, and the `p` as a pointer to the head of the packed list.

(This requires adjustment to the routines that called `va_arg()` to
consider consulting the packed array instead of the va_list.  The code
involved requires significant review, including the question of if
it is worth it to try to avoid building a temporary array up-front.
Pending that review, the changes to deal with `packed` vs `vaptr` are
somewhat ad-hoc, vs. introducing yet another layer of abstraction.)

To make the conversion of C++ types extensible, this makes use of
overloading so that global overloads of `to_rebarg` are used by the
template to construct transient values (e.g. carrying the rebR() flag)
that will be passed to the feed.