Formalize "non macro api call" in libRebol
The use of C99 macros to inject `rebEND` at the end of the call
conflicts with cases where one wants to use #ifdefs in the middle
of the call itself.  Previously, getting around this was done by
calling the internal implementation detail routines named _inline.
However, that no longer applies in the C++ build as the variadic
wrapper does the work itself.

This changes to a formalization based on LIBREBOL_NOMACRO() which
takes an API function and gives the variant of it that does not use
macros.  Whether it's the best name or not, it's a better way of
capturing the intent.  In the C++ build it simply is a no-op and
returns the same API passed in.