Improve + refactor rebSpell/rebSpellWide/rebBytes
The APIs for extracting string and byte data came in two forms: one
that allocated the appropriate sized buffer for you, and another API
that could be called with a buffer and a size and get the actual
size back.  Implementing the allocating version was done by passing
a null buffer in to get the size, and then making a second call.

The functions had drifted somewhat, so that the rebBytes() function
supported more types than the rebBytesInto() function.  Also, the
functions were implemented on top of the variadic forms...which meant
that quoting had to be used on the second call, and this incurred
cost from making variadic feeds twice.

This makes the XXX and XXXInto forms based on a common non-variadic
function, improving performance and making the capabilities equivalent.