Use packed argument form for JavaScript interop
C uses va_list to pass variadic parameters because that is the only
way to do it.  In general, there is no standard for how va_list is
implemented...and they cannot be created programmatically in C.
So that makes them a bit unpredictable.

The JavaScript build only uses clang in a predictable configuration.
So it was possible to hard-code a parallel way of building va_lists in
JavaScript with the Wasm HEAP32 to what the compiler expected.  But
since interop with C++ required a standard way of passing an array
of packed pointers, it makes more sense to use that for JavaScript too.

It's not that much different...but slightly clearer and saves on one
pointer's worth of memory and initialization per call.