Get rid of rebEnterApi_internal() as an API
In order to avoid having to type in boilerplate code on each API
for any common entry code, a function was defined that would be put
in the external wrappers automatically.

While expedient, this reduced efficiency by forcing an additional
function call of overhead for something that could be inlined in the
library code itself.  It also complicated the wrapping code, and made
the header file look "junkier" to anyone browsing through it.

Given that the API has been on a fairly stable track to have limited
entry points, it's not that big a deal to add ENTER_API (certainly
no worse than having to put a macro inclusion at the top of every
native implementation).  This goes ahead and does the simplification.