Implement experimental -1-SLASH- alias for /
It's very desirable to have `/`, `/foo`, `/foo/`, `/foo/(bar)` etc. be
instances of the same datatype of PATH!.  In this scheme, `/` would act
like a "root path" and be achieved with `to path! [_ _]`.

But with limited ASCII symbols, there is strong demand for `/` to be
able to act like division in evaluative contexts, or to be overrideable
for other things in a way not too dissimilar from `+` or `-`.

This is a very experimental compromise, to make `/` a cell whose
VAL_TYPE() is REB_PATH, but whose CELL_KIND() is REB_WORD with the
special spelling `-1-SLASH-`.  Binding mechanics and evaluator behavior
are based on this unusual name.  But when inspected by the user, it
appears to be a PATH! with 2 blanks.

Little testing has been done on this, so use caution.