Scrap idea of cell lifetime bits in cell itself
In the early introduction of REB_FRAME, the idea of making a varlist
per function invocation seemed very costly in terms of the GC.  This
led to thinking about ways to avoid needing to manage varlists, so
they could be freed when the function finished instead of being kept
alive.

The only varlists that would have to be kept would be those whose
pointers "escaped" to be used as the bindings of arguments or locals
in the frame...or if the FRAME! itself were referenced as a value.
This would not typically happen for natives (in non-debug scenarios
which may inspect and refer to native args).  But any client code
that passed a WORD! or array containing a WORD! bound to the frame
as a local would count as a "leak" and require management.

This led to thought experiments about how to determine if a leak would
be to cells that would naturally be cleared up before the function
finished...e.g. passing cells to deeper stack levels.  Primitive means
of marking things like the data stack as "transient" or flagging cells
as being "stacklike" were placeholders for more elaborate ideas related
to this optimization.

(Some thought was that the extra header bits on 64-bit platforms might
be used to hold something like a stack level index, and this could
help know whether a cell was being moved into a stack level that would
outlive its binding or not.)

Other objectives--such as stackless evaluation--made such optimization
likely impossible.  For the sake of reducing the complexity, this
eliminates the related code...which reclaims two node bits, that might
be useful for other more viable optimization strategies.