Return support for value in TRAP and CATCH
In order to make TRAP and CATCH "cleaner" primitives that did not
conflate evaluative results with things that were actually "trapped"
(due to FAIL) or "caught" (due to THROW), they returned null when an
actual failure or throw did not occur.  This adds an extra optional
return value for the result of the evaluation.

For example:

    >> [e v]: trap [1 + 2]
    ; null

    >> e
    ; null

    >> v
    == 3