Introduce "unrelativize" routine
Relative value cells are those that are in a function's body, and are
tied to a function instance in order to look up the values of words.
Specific cells are those that don't live in a function's body, which
may be ones that were from a function's body but have been
"derelativized" to join persistently with a function instance.

Ren-C's type system enforces distinct handling of relative and specific
values, letting only specific values leak to the user.  This causes
some issues for error messages that need to be emitted by lower-level
mechanics which are not passed specifiers...because they cannot use
the raw material without connecting it to a function.

Several cases of low-level routines thus become complicated by forcing
the caller to pass in a specifier--not because of the informational
needs of the routine itself, but because of the error message it may
wind up delivering.  This is good if you want the error message to
contain correct bindings in case its raw material is queried by the
user (as opposed to just displayed).  But the cost in complexity can
be significant for some errors.

This introduces the option of being able to "Unrelativize" a value,
which binds relative values to an *expired* function context that is
generated on the fly if necessary.  If possible this should be avoided,
but when a low-level context would be too annoying to implement another
way then it can be used.

Also changes FAIL_IF_READ_ONLY for a value to ENSURE_MUTABLE, so the
value can be passed through (following the ENSURE native protocol).