Differentiate mutable and immutable series access
Historically, the checking of protection bits like those from PROTECT
were done on an "honor system".  C code that did mutations would have
to remember to check the bits for protection for modifying operations.
Failure to do so would be a bug, and there were many cases of this.

As CONST and MUTABLE have become core features (vs. a "fringe" feature
like PROTECT that wasn't really used), the importance of doing checks
of mutability increases.  This commit formalizes the checks in the
interpreter layered on the `const` feature of the C and C++ compiler.

Historically a barrier to leveraging the `const` state on
REBSER/REBARR/REBMAP/etc. was that there were many accessor routines
which needed to work on both mutable and immutable series.  But due to
the absence of C overloading it was not possible to give these routines
the same name.  This resulted in combinatorics that were not desirable,
making generic macros that would work on both mutable and immutable
series impossible.

This uses a trick which compromises some const checking in C builds,
leaving full validation to the C++ build.  However, the C code still
compiles.  This trick leverages the `const_if_c` flag.