Rework DEBUG_CHECK_CASTS, incorporate `const`
This unifies the control of extended cast checking with `cast()` to
go with the checking of SER(), ARR(), etc.

It also includes various fixes and upgrades to the SER(), ARR() etc.
macros...including some handling for `const`...such that a SER() on
a const pointer can only give back a const REBSER*.  This is used on
the stackless branch in checking the CONST/MUTABLE/PROTECT status
rigorously.