`const_if_c` definition
This creates a notion of `const_if_c`, which can be used to make a
routine permissive on a const input giving mutable output...but then
a C++ build can augment that with another signature that has `const`
and returns const output.  This helps avoid combinatorial explosions
of names due to C not being able to overload, leaving the full const
checking to the C++ build.