C-based heuristic workaround for strerror_r
The interface for `strerror_r` is a historical debacle, because the
GNU compiler defined it differently from the POSIX standard but used
the same name.  This makes it a chronic weak-link when compiling in
different configurations.

Detecting the difference with C++ involves overloading and does not
work in pure C (this is used by Chromium, for instance).  In C, one can
use older interfaces like accessing the internal string table directly,
but this could miss some error messages that don't use the table.

Historically the way of telling what strerror_r definition is in effect
in C involves a complex macro suggested by the GNU documentation:

  The XSI-compliant version of strerror_r() is provided if:
  (_POSIX_C_SOURCE >= 200112L || _XOPEN_SOURCE >= 600) && ! _GNU_SOURCE
  Otherwise, the GNU-specific version is provided.

Early Ren-C took this for granted.  But it appears modern GCC 9 is
giving the POSIX version even if _GNU_SOURCE is defined.  Also, other
libraries like musl sound as if they don't get this tapdance correct.

This commit takes a new approach of casting the `int` or `char*` result
to an `intptr_t`.  Then based on looking at values of that intptr_t
in coordination with errno changes and mutations to the passed in
buffer, it decides what to do.  This will likely reduce pain in the
build process related to this issue.